% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/screengeoVL.R
\name{AreaHull}
\alias{AreaHull}
\title{Calculate area of the complex hull envelopping a visual stimulus}
\usage{
AreaHull(x, y)
}
\arguments{
\item{x}{vector of x coordinates of pixels}

\item{y}{vector of y coordinates of pixels}
}
\value{
a list with the points component containing the points defining the hull and a scalar component containing the area of the hull.
}
\description{
Calculate area of the complex hull envelopping a visual stimulus
}

