% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/screengeoVL.R
\name{DVA_VL}
\alias{DVA_VL}
\title{Calculate degrees of visual angle}
\usage{
DVA_VL(stimX, stimY, dist, distUnit = "cm", scrX, scrY, scrUnit = "mm",
  resX, resY)
}
\arguments{
\item{stimX}{Stimulus width in pixels}

\item{stimY}{Stimulus height in pixels}

\item{dist}{Distance from screen to user}

\item{distUnit}{Measurement unit for dist}

\item{scrX}{screen width}

\item{scrY}{Screen height}

\item{scrUnit}{Measurement unit for scrX and scrY}

\item{resX}{Screen resolution (horizontal)}

\item{resY}{Screen resolution (vertical)}
}
\value{
A list with the visual angles in radian and in degrees
}
\description{
Calculate degrees of visual angle
}

